requirements.txt
annotated-types
anyio
fastapi
idna
packaging
Pillow
pydantic
pydantic_core
pytesseract
sniffio
starlette
typing_extensions
uvicorn
python-multipart
cv2